% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{findLocalFvalMax}
\alias{findLocalFvalMax}
\title{Determines the local maxes of F-test values}
\usage{
findLocalFvalMax(fval, k, cutoff)
}
\arguments{
\item{fval}{a \code{vector} containing the value sof the f-test.}

\item{k}{an \code{integer} indicating the number of tapers used to calculate the F-values.}

\item{cutoff}{a value in [0, 1] representing the significance level.}
}
\description{
Finds the local max of the multitaper harmonic F-test by checking the neighbours.
}
