% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{findLocalMscMax}
\alias{findLocalMscMax}
\title{Determines the local maxes of MSC}
\usage{
findLocalMscMax(msc, k, cutoff)
}
\arguments{
\item{msc}{the normal transformed magnitude square coherence (see \link{mscQTransform})}

\item{k}{an \code{integer} indicating the number of tapers used to calculate the MSC}

\item{cutoff}{a value in [0, 1] representing the significance level.}
}
\description{
Finds the local max of the normal transformed MSC by checking the neighbours.
}
