% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{msc2norm}
\alias{msc2norm}
\title{Normal Transform for Magnitude Squared Coherence}
\usage{
msc2norm(msc, dof)
}
\arguments{
\item{msc}{A vector containing the magnitude squared coherence.}

\item{dof}{Degrees of freedom of the msc}
}
\description{
Converts the MSC to a Gaussian distribution
}
\details{
This code was provided by Emily Somerset.
}
