% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{specPredict}
\alias{specPredict}
\title{Spectrum prediction
Predicts the spectrum based on an estimated transfer function and new data provided.}
\usage{
specPredict(H, Hinfo, info, d2)
}
\arguments{
\item{H}{The H returned by tf().}

\item{Hinfo}{The Hinfo returned by tf().}

\item{info}{The info returned by tf().}

\item{d2}{A \code{data.frame} containing the new data.  Must have the same column names as the original 
data used in the transfer function estimation.}
}
\description{
Spectrum prediction
Predicts the spectrum based on an estimated transfer function and new data provided.
}
