% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{tf}
\alias{tf}
\title{Estimates the transfer functions}
\usage{
tf(d1, d2, ndata = length(d1), ndata2 = length(d2), blockSize = ndata,
  blockSize2 = blockSize, overlap = 0, dt = 1, dt2 = dt, nw = 4,
  nw2 = NULL, k = 7, nFFT = NULL, nFFT2 = nFFT, freqRange = NULL,
  maxFreqOffset = 0, nOff = -1, sigLevel = 0.99, name1 = "d1",
  name2 = "d2")
}
\arguments{
\item{d1}{A real vector containing the response.}

\item{d2}{A \code{data.frame} whose columns contain the predictors.}
}
\description{
Estimates a transfer function using frequency offsets between a response and multiple inputs
}
