\name{transfer2-package}
\alias{transfer2-package}
\alias{transfer2}
\docType{package}
\title{
\packageTitle{transfer2}
}
\description{
\packageDescription{transfer2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{transfer2}
\packageIndices{transfer2}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{transfer2}

Maintainer: \packageMaintainer{transfer2}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
