% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing2.R
\name{tstEigenCoef}
\alias{tstEigenCoef}
\title{Taper Testing}
\usage{
tstEigenCoef()
}
\description{
Should get the same answer as from tapering in R
}
\details{
Looks good - differences are a result of the implementation of the FFT's between 
R and fftpack5 that I'm using in Fortran - differences on the order of 1e-14 at most.
}
