% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing2.R
\name{tstWeightedEigenCoef}
\alias{tstWeightedEigenCoef}
\title{Adaptive Eigencoefficient Testing}
\usage{
tstWeightedEigenCoef(d1 = NULL, d2 = NULL)
}
\description{
Should get the same answer as from multitaper in R
}
\details{
Looks good - differences are a result of the implementation of the FFT's between 
R and fftpack5 that I'm using in Fortran - differences on the order of 1e-14 at most.
}
