% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing2.R
\name{tstWorking}
\alias{tstWorking}
\title{Tests to make sure that Fortran calls are working (again)}
\usage{
tstWorking()
}
\description{
Tests to make sure that Fortran calls are working (again)
}
\details{
The answers are different due to where sqrt(dt) (?) gets taken into account
}
