% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zFilter}
\alias{zFilter}
\title{Convolve Complex-Valued Impulse Response}
\usage{
zFilter(x, filter, realPart = TRUE)
}
\arguments{
\item{x}{A \code{vector} on which to apply the impulse response.}

\item{filter}{A \code{vector} the impulse response, must be odd-length (for now).}
}
\description{
Convolve Complex-Valued Impulse Response
}
\details{
Taken from transfer package (Aaron and I wrote that).
This only accepts two-sided filtering for now.
}
