% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherence}
\alias{coherence}
\title{Calculate the multitaper coherence estimate}
\usage{
coherence(d1, d2, ndata = length(d1), blockSize = ndata, overlap = 0,
  dt = 1, nw = 4, k = 7, nFFT = NULL, freqRange = NULL,
  maxFreqOffset = 0, calcType = 1, forward = 1, name1 = "d1",
  name2 = "d2")
}
\arguments{
\item{d1}{A \code{numeric} vector containing the first series.}

\item{d2}{A \code{numeric} vector containing the second series (same length as \code{d1}).}

\item{ndata}{The total length of d1 or d2.}

\item{blockSize}{The length of a single block to use (if blocking)}

\item{overlap}{A \code{numeric} value in the range [0, 1) indicating the proporation of 
overlap between neighbouring blocks.}

\item{dt}{The sampling rate in seconds.}

\item{nw}{time-bandwidth parameter for multitaper}

\item{k}{number of tapers to use (k < 2*nw)}

\item{nFFT}{the number of frequency bins to use (nFFT > 2*ndata)}

\item{freqRange}{A vector with 2 elements containing the start and end frequencies (in Hz) 
over which to calculate the coherence.}

\item{maxFreqOffset}{Every pair of frequencies between f1 (series 1) 
and f1 +/- maxFreqOffset (series 2) will be calculated (f1 + maxFreqOffset < nyquist)}

\item{calcType}{An \code{integer} value indicating how averaging over blocks should 
be performed:
1 - calculate the MSC on each block, then average;
2 - calculate the cross and auto spectra on each block, average each quantity 
across blocks, then calculate the coherency;
3 - calculate the coherency on each block, then average
4 - returns the minimum MSC across blocks}

\item{forward}{An \code{integer} indicating whether the forward (1) or reverse (0) coherence
should be calculated.}

\item{name1}{a \code{character} string giving the name of the first data set used.}

\item{name2}{a \code{character} string giving the name of the second data set used.}
}
\description{
Calculates the magnitude squared coherence estimate
}
