% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{mscQTransform}
\alias{mscQTransform}
\title{Transform coherence to Guassian distribution}
\usage{
mscQTransform(c, k, msc = TRUE)
}
\arguments{
\item{c}{a real-valued vector of the coherence}

\item{k}{the number of tapers used}
}
\description{
Converts estimates of MSC to an almost Gaussian distribution
}
