% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing2.R
\name{tstWeightedEigenCoef}
\alias{tstWeightedEigenCoef}
\title{Adaptive Eigencoefficient Testing}
\usage{
tstWeightedEigenCoef(d1 = NULL, d2 = NULL)
}
\description{
Should get the same answer as from multitaper in R
}
\details{
Less good - differences on the order of 1e-7 ...
}
